Safari-609.1.11.1/LayoutTests/fast/css/getComputedStyle/computed-style-page-break-inside-expected.txt

Test that page-break-inside property is not inherited

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(parent).getPropertyCSSValue('page-break-inside').getStringValue() is "avoid"
PASS window.getComputedStyle(child).getPropertyCSSValue('page-break-inside').getStringValue() is "auto"
PASS successfullyParsed is true

TEST COMPLETE
